run update
Sorry for the "death by papercuts" commits. I use a linting/code maintenance tool called `update` that ensures that code style and configuration settings meet my standards and conventions. I recently refactored `update` so the first time I run it in a project it will typically find a lot of things to fix.