Removing "drop-in" from readme
Thank you very much for creating this project! Opening a pr since it seems easier to discuss here.

Maybe we have different definitions, but I expect a drop in replacement to be just a change in the require. That is certainly not the case.

In particular I was trying to replace minimatch in glob with micromatch (since the performance can be horrific), but it's pretty involved.
Reference: https://github.com/isaacs/node-glob/blob/master/glob.js#L45