Fix: Put mm into bash mode for bash tests
It would appear that a section of the bash tests were being run without
micromatch being given the `bash` option.

Note: Supplying this option causes some of the tests to fail. In
particular, it would appear that in bash mode `*` will match into
directories like `**`, for example `*/c` will match `a/b/c`.